add cargo update
